added graphics flag

+++ sim/hw2/main.go
@@ -7,13 +7,15 @@ import (
 	"github.com/stanim/svgof/svgo2f"
 	"os"
 	"log"
+	"flag"
 )
 
+var MINDIFF float64 = 0.000001
+var DARTS int = 30000
+
 const (
-	DARTS = 30000
 	LEFTEND = 2
 	RIGHTEND = 4
-	MINDIFF = 0.000001
 	LOWERBOUND = 1
 	AREA = 1
 )
@@ -61,6 +63,12 @@ func f(x float64) float64 {
 }
 
 func main() {
+	graphics := false
+	flag.BoolVar(&graphics, "graphics", graphics, "Toggle graphics")
+	flag.Float64Var(&MINDIFF, "mindiff", MINDIFF, "Difference until can quit")
+	flag.IntVar(&DARTS, "darts", DARTS, "Darts used")
+	flag.Parse()
+
 	bounds := &Bounds { Left: LEFTEND, Right: RIGHTEND }
 	for math.Abs(bounds.Right-bounds.Left) >= MINDIFF {
 		midpoint := bounds.Midpoint()
@@ -81,43 +89,45 @@ func main() {
 			}
 		}
 
-		area := AREA - ((underCurve/DARTS) * ((midpoint-LOWERBOUND) * f(LOWERBOUND)))
+		area := AREA - ((underCurve/float64(DARTS)) * ((midpoint-LOWERBOUND) * f(LOWERBOUND)))
 		bounds = bounds.Enclose(area)
 	}
 	fmt.Println(bounds);
 
-	width := 800.0
-	height := 800.0
-	hd := (height-10)/f(LOWERBOUND)
-	wd := (width-10)/(RIGHTEND-LOWERBOUND)
-	count := 0
-	dirName := "graphs"
-	if err := os.RemoveAll(dirName); err != nil {
-		log.Fatal(err)
-	}
-	if err := os.Mkdir(dirName, 0777); err != nil {
-			log.Fatal(err)
-	}
-	for b := bounds.last; b != nil; b = b.last {
-		file, err := os.Create(fmt.Sprintf("%s/graph_%.3d.svg", dirName, count))
-		if err != nil {
+	if graphics {
+		width := 800.0
+		height := 800.0
+		hd := (height-10)/f(LOWERBOUND)
+		wd := (width-10)/(RIGHTEND-LOWERBOUND)
+		count := 0
+		dirName := "graphs"
+		if err := os.RemoveAll(dirName); err != nil {
 			log.Fatal(err)
 		}
-		canvas := svg.New(file)
-		canvas.Start(width, height)
-		canvas.Rect(0, 0, width, height, `fill="beige"`)
-		for _, d := range b.Darts {
-			if d.y < f(d.x) {
-				if d.x > b.Left {
-					canvas.Circle(((d.x - LOWERBOUND)*wd)+5, height - ((d.y*hd)+5), 2, `fill="red"`)
-				} else {
-					canvas.Circle(((d.x - LOWERBOUND)*wd)+5, height - ((d.y*hd)+5), 2, `fill="blue"`)
-				}
-			} else {
-				canvas.Circle(((d.x - LOWERBOUND)*wd)+5, height - ((d.y*hd)+5), 2)
+		if err := os.Mkdir(dirName, 0777); err != nil {
+				log.Fatal(err)
+		}
+		for b := bounds.last; b != nil; b = b.last {
+			file, err := os.Create(fmt.Sprintf("%s/graph_%.3d.svg", dirName, count))
+			if err != nil {
+				log.Fatal(err)
 			}
+			canvas := svg.New(file)
+			canvas.Start(width, height)
+			canvas.Rect(0, 0, width, height, `fill="beige"`)
+			for _, d := range b.Darts {
+				if d.y < f(d.x) {
+					if d.x > b.Left {
+						canvas.Circle(((d.x - LOWERBOUND)*wd)+5, height - ((d.y*hd)+5), 2, `fill="red"`)
+					} else {
+						canvas.Circle(((d.x - LOWERBOUND)*wd)+5, height - ((d.y*hd)+5), 2, `fill="blue"`)
+					}
+				} else {
+					canvas.Circle(((d.x - LOWERBOUND)*wd)+5, height - ((d.y*hd)+5), 2)
+				}
+			}
+			canvas.End()
+			count++
 		}
-		canvas.End()
-		count++
 	}
 }
